✨ implement TeamFormPaper
