+++ src/components/domains/team/TeamFormPaper/index.ts
@@ -1,0 +1,1 @@
+export { TeamFormPaper } from './TeamFormPaper';
